backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.13.0
requests==2.31.0
pyyaml==6.0.2
pytest==7.4.4
pytest-asyncio==0.23.3
